+++ lib/persistence.dart
@@ -1,0 +1,5 @@
+/// The Flutter SDK for Aptabase, a privacy-first and simple analytics platform for apps.
+library aptabase_persistence;
+
+export 'src/offline_logic/services/events_service.dart';
+export 'src/offline_logic/event_offline.dart';
