+++ lib/src/offline_logic/connectivity_checker.dart
