+++ lib/src/offline_logic/services/endpoint_base_abstract.dart
