+++ test/event_persist_test.dart
@@ -1,5 +1,5 @@
-import 'package:aptabase_flutter/event.dart';
-import 'package:aptabase_flutter/event_service.dart';
+import 'package:aptabase_flutter/src/offline_logic/event_offline.dart';
+import 'package:aptabase_flutter/src/offline_logic/services/events_service.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sembast/sembast.dart';
 import 'package:sembast/sembast_memory.dart';
@@ -10,25 +10,26 @@ void main() async {
   final db = await databaseFactoryMemory.openDatabase('test.db');
   test('save and get', () async {
     final dbStore = intMapStoreFactory.store('aptabase');
-    final int key = await dbStore.add(db, Event.dummy.toMap());
-    final int key2 = await dbStore.add(db, Event.dummy2.toMap());
+    await dbStore.add(db, EventOffline.dummy.toMap());
+    await dbStore.add(db, EventOffline.dummy2.toMap());
     // Build our app and trigger a frame.
     final recordSnapshot = await dbStore.find(db);
-    final events = recordSnapshot.map((e) => Event.fromMap(e.value)).toList();
+    final events =
+        recordSnapshot.map((e) => EventOffline.fromMap(e.value)).toList();
     expect(events.length, 2);
     // final events = getAllPersistedEvents(prefs);
-    expect(events.last, Event.dummy2);
+    expect(events.last, EventOffline.dummy2);
     await dbStore.delete(db);
   });
 
-  test('test prevent from saving events if already too many', () async {
-    final oneThousandEvents =
-        List<Event>.generate(1000, (index) => const Event('eventName'));
+  test('test prevent from saving events if already 1000', () async {
+    final oneThousandEvents = List<EventOffline>.generate(
+        1000, (index) => const EventOffline('eventName'));
     final service = TestEventService(db);
     for (final e in oneThousandEvents) {
       await service.addEvent.request(e);
     }
-    final isPersisted = await service.addEvent.request(Event.dummy);
+    final isPersisted = await service.addEvent.request(EventOffline.dummy);
     expect(isPersisted, false);
   });
 }
