--- test/mock_test.dart
@@ -1,10 +1,0 @@
-import 'dart:developer';
-
-import 'package:aptabase_flutter/aptabase_flutter.dart';
-import 'package:aptabase_flutter/event.dart';
-import 'package:aptabase_flutter/event_service.dart';
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  test('test that event saved is indeed sent', () async {});
-}
